(refactor): redundant class methods in CreateSubmissionPDFAction classes to abstract class
