+++ src/Contracts/AbstractCreateSubmissionPDFAction.php
@@ -9,9 +9,14 @@ use OWC\Zaaksysteem\Entities\Zaak;
 use OWC\Zaaksysteem\Entities\Zaakinformatieobject;
 use OWC\Zaaksysteem\GravityForms\GravityPdfSettings;
 use OWC\Zaaksysteem\Resolvers\ContainerResolver;
+use OWC\Zaaksysteem\Traits\CheckURL;
+use OWC\Zaaksysteem\Traits\InformationObject;
 
 abstract class AbstractCreateSubmissionPDFAction
 {
+    use CheckURL;
+    use InformationObject;
+
     public const CLIENT_NAME = '';
     public const FORM_SETTING_SUPPLIER_KEY = '';
 
@@ -30,7 +35,7 @@ abstract class AbstractCreateSubmissionPDFAction
         $this->gravityPdfSettings = new GravityPdfSettings($entry, $form);
     }
 
-    abstract protected function getSubmissionArgsPDF(): array;
+    abstract public function addSubmissionPDF(): ?Zaakinformatieobject;
 
     public function getInformationObjectType(): string
     {
@@ -56,5 +61,61 @@ abstract class AbstractCreateSubmissionPDFAction
         }
 
         return $this->client->zaakinformatieobjecten()->create(new Zaakinformatieobject($pdf->toArray(), $this->client->getClientName(), $this->client->getClientNamePretty())); // What to do when this one fails?
+    }
+
+    /**
+     * Get the generated PDF of the submission.
+     */
+    protected function getSubmissionArgsPDF(): array
+    {
+        if (! class_exists('GPDFAPI')) {
+            return [];
+        }
+
+        if (! $this->gravityPdfSettings->pdfFormSettingIsActive()) {
+            return [];
+        }
+
+        $pdfURL = $this->gravityPdfSettings->pdfURL();
+
+        if (empty($pdfURL) || ! $this->checkURL($pdfURL)) {
+            return [];
+        }
+
+        // Enable the public access setting so the args can be prepared.
+        $this->gravityPdfSettings->updatePublicAccessSettingPDF('enable');
+
+        $args = $this->prepareFormSubmissionArgsPDF('Aanvraag - eFormulier', $pdfURL);
+
+        // Disable the public access setting again so the PDF stays protected.
+        $this->gravityPdfSettings->updatePublicAccessSettingPDF();
+
+        return $args;
+    }
+
+    public function prepareFormSubmissionArgsPDF(string $fileName, string $pdfURL): array
+    {
+        $informationObectType = $this->getInformationObjectType();
+
+        if (empty($informationObectType)) {
+            return [];
+        }
+
+        $args = [];
+        $args['titel'] = $fileName;
+        $args['formaat'] = $this->getContentType($pdfURL);
+        $args['bestandsnaam'] = sprintf('%s.pdf', \sanitize_title($fileName));
+        $args['bestandsomvang'] = (int) $this->getContentLength($pdfURL);
+        $args['inhoud'] = $this->informationObjectToBase64($pdfURL);
+        $args['vertrouwelijkheidaanduiding'] = 'vertrouwelijk';
+        $args['auteur'] = 'OWC';
+        $args['status'] = 'gearchiveerd';
+        $args['taal'] = 'nld';
+        $args['versie'] = 1;
+        $args['bronorganisatie'] = ContainerResolver::make()->rsin();
+        $args['creatiedatum'] = date('Y-m-d');
+        $args['informatieobjecttype'] = $informationObectType;
+
+        return $args;
     }
 }
